Fix integer passed to make

+++ pkg/ebpf/tracer_windows.go
@@ -149,7 +149,7 @@ func (t *Tracer) resizeBuffer(compareSize int, buffer []network.ConnectionStats)
 		return make([]network.ConnectionStats, 0, cap(buffer)*2)
 	} else if compareSize <= cap(buffer)/2 {
 		// Take the max of buffer/2 and compareSize to limit future array resizes
-		return make([]network.ConnectionStats, 0, math.Max(float64(cap(buffer)/2), float64(compareSize)))
+		return make([]network.ConnectionStats, 0, int(math.Max(float64(cap(buffer)/2), float64(compareSize))))
 	}
 	return buffer
 }
